bringing the whole project into a clean structute
