--- src/components/Profile/Profile.css
@@ -1,110 +1,0 @@
-.profile-container {
-  margin-left: 20rem;
-  padding: 0rem 1.5rem;
-}
-
-.profile-card {
-  border-radius: 10px;
-  background: #1e2a47;
-  margin-bottom: 1.5rem;
-  padding: 1rem;
-  height: auto;
-  color: #fff;
-}
-
-.profile-card .top-container {
-  display: flex;
-  gap: 1rem;
-}
-
-.profile-card .profile-photo {
-  height: 6.1rem;
-  width: 6.1rem;
-  flex-shrink: 0;
-}
-
-.profile-photo img {
-  width: 100%;
-  height: 100%;
-  border-radius: 50%;
-}
-
-.profile-details {
-  width: 44rem;
-  overflow: hidden;
-}
-
-.profile-details p {
-  font-size: 14px;
-}
-
-.profile-details .skills-container {
-  margin-top: 1rem;
-  display: flex;
-  gap: 1rem;
-  height: 35px;
-  overflow: hidden;
-}
-
-.skills-container:hover {
-  overflow-x: scroll;
-  scroll-behavior: auto;
-}
-
-.skills-container::-webkit-scrollbar {
-  width: 10px;
-  height: 5px;
-}
-
-.skills-container::-webkit-scrollbar-track {
-  background: #1e2a47;
-}
-
-.skills-container::-webkit-scrollbar-thumb {
-  background: rgba(20, 29, 47, 0.5);
-  border-radius: 5px;
-}
-
-.skills-container .skill {
-  background-color: rgba(20, 29, 47, 1);
-  padding: 0.25rem 0.5rem;
-  border-radius: 5px;
-  font-size: 0.875rem;
-  white-space: nowrap;
-  height: 30px;
-  cursor: default;
-}
-
-.profile-link a {
-  text-decoration: none;
-  color: #00a6fb;
-}
-
-.bottom-container {
-  margin-top: 1rem;
-}
-
-.bottom-container .social {
-  display: flex;
-  margin-top: 0.5rem;
-}
-
-.social .github,
-.social .twitter,
-.social .linkedin {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-  margin-right: 2rem;
-}
-
-.social-link {
-  text-decoration: none;
-  color: #fff;
-}
-
-@media (max-width: 480px) {
-  .skills-container .skill {
-    height: 2rem;
-  }
-}
