+++ src/components/index.js
@@ -1,0 +1,6 @@
+import Search from "./Search/Search";
+import Sidebar from "./Sidebar/Sidebar";
+import Profile from "./Profile/Profile";
+import Card from "./Card/Card";
+import NoResultFound from "./NoResultFound/NoResultFound";
+export { Search, Sidebar, Profile, NoResultFound, Card };
